changed angle to int

+++ src/main/java/be/ugent/objprog/ugentopoly/gameBoard/bars/Bar.java
@@ -3,5 +3,5 @@ package be.ugent.objprog.ugentopoly.gameBoard.bars;
 public interface Bar {
     void populate();
 
-    void applyRotation(double angle);
+    void applyRotation(int angle);
 }
